Add import for custom user model class in "Referencing the User model" example.
If you don't import the user class when adding a foreign key to a model, it will raise a "ValueError". The docs do not make this clear from the example above. 

Also: this error does not occur when using the testserver, but _will_ occur on a production server such as gunicorn.